@@ -1,9 +1,9 @@
-import { Directive, TemplateRef } from '@angular/core';
+import { Directive, inject, TemplateRef } from '@angular/core';
 
 @Directive({
   selector: '[ngsSnippetResult]',
   standalone: true,
 })
 export class SnippetResultDirective {
-  constructor(public templateRef: TemplateRef<any>) {}
+  public templateRef: TemplateRef<any> = inject(TemplateRef);
 }
