@@ -1,4 +1,10 @@
-import { Directive, ElementRef, HostBinding, OnDestroy } from '@angular/core';
+import {
+  Directive,
+  ElementRef,
+  HostBinding,
+  inject,
+  OnDestroy,
+} from '@angular/core';
 
 import { NavigationFocusService } from './navigation-focus.service';
 
@@ -11,20 +17,22 @@ export class NavigationFocusDirective implements OnDestroy {
   @HostBinding('tabindex') readonly tabindex = '-1';
   @HostBinding('style.outline') readonly outline = 'none';
 
-  constructor(
-    private el: ElementRef,
-    private navigationFocusService: NavigationFocusService
-  ) {
-    if (!el.nativeElement.id) {
-      el.nativeElement.id = `skip-link-target-${uid++}`;
+  private el: ElementRef = inject(ElementRef);
+  private navigationFocusService: NavigationFocusService = inject(
+    NavigationFocusService,
+  );
+
+  constructor() {
+    if (!this.el.nativeElement.id) {
+      this.el.nativeElement.id = `skip-link-target-${uid++}`;
     }
-    this.navigationFocusService.requestFocusOnNavigation(el.nativeElement);
-    this.navigationFocusService.requestSkipLinkFocus(el.nativeElement);
+    this.navigationFocusService.requestFocusOnNavigation(this.el.nativeElement);
+    this.navigationFocusService.requestSkipLinkFocus(this.el.nativeElement);
   }
 
   ngOnDestroy() {
     this.navigationFocusService.relinquishFocusOnNavigation(
-      this.el.nativeElement
+      this.el.nativeElement,
     );
     this.navigationFocusService.relinquishSkipLinkFocus(this.el.nativeElement);
   }
