@@ -1,4 +1,4 @@
-import { Directive, Input, TemplateRef } from '@angular/core';
+import { Directive, inject, Input, TemplateRef } from '@angular/core';
 
 @Directive({
   selector: '[ngsRoutedTab]',
@@ -8,5 +8,5 @@ export class RoutedTabDirective {
   @Input('ngsRoutedTab') routeName: string;
   @Input('ngsRoutedTabLabel') label: string;
 
-  constructor(public templateRef: TemplateRef<any>) {}
+  public templateRef: TemplateRef<any> = inject(TemplateRef);
 }
