@@ -1,5 +1,5 @@
 /* eslint-disable @angular-eslint/directive-selector */
-import { Directive, OnDestroy } from '@angular/core';
+import { Directive, inject, OnDestroy } from '@angular/core';
 import { MatListItem } from '@angular/material/list';
 import { RouterLinkActive } from '@angular/router';
 
@@ -12,11 +12,13 @@ import { Subscription } from 'rxjs';
 export class RouterActivatedMatListItemDirective implements OnDestroy {
   private subs = new Subscription();
 
-  constructor(matListItem: MatListItem, routerLinkActive: RouterLinkActive) {
+  constructor() {
+    const matListItem = inject(MatListItem);
+    const routerLinkActive = inject(RouterLinkActive);
     this.subs.add(
       routerLinkActive.isActiveChange.subscribe((isActive) => {
         matListItem.activated = isActive;
-      })
+      }),
     );
   }
 
