@@ -1,4 +1,4 @@
-import { Directive, ElementRef, Input, OnChanges } from '@angular/core';
+import { Directive, ElementRef, inject, Input, OnChanges } from '@angular/core';
 
 @Directive({
   selector: '[ngsFocusable]',
@@ -8,7 +8,7 @@ export class FocusableDirective implements OnChanges {
   @Input('ngsFocusable') focusable: boolean;
   @Input() ngsFocusableSelector: string;
 
-  constructor(private elRef: ElementRef) {}
+  private elRef = inject(ElementRef);
 
   ngOnChanges() {
     const el: HTMLElement = this.elRef.nativeElement;
