@@ -9,40 +9,40 @@ export class FakeMediaQueryList implements MediaQueryList {
   media: string;
   onchange: ((this: MediaQueryList, ev: MediaQueryListEvent) => any) | null;
   addListener(
-    _callback: ((this: MediaQueryList, ev: MediaQueryListEvent) => any) | null
+    _callback: ((this: MediaQueryList, ev: MediaQueryListEvent) => any) | null,
   ): void {}
   removeListener(
-    _callback: ((this: MediaQueryList, ev: MediaQueryListEvent) => any) | null
+    _callback: ((this: MediaQueryList, ev: MediaQueryListEvent) => any) | null,
   ): void {}
   addEventListener<K extends 'change'>(
     type: K,
     listener: (this: MediaQueryList, ev: MediaQueryListEventMap[K]) => any,
-    options?: boolean | AddEventListenerOptions | undefined
+    options?: boolean | AddEventListenerOptions | undefined,
   ): void;
   addEventListener(
     type: string,
     listener: EventListenerOrEventListenerObject,
-    options?: boolean | AddEventListenerOptions | undefined
+    options?: boolean | AddEventListenerOptions | undefined,
   ): void;
   addEventListener(
     _type: unknown,
     _listener: unknown,
-    _options?: unknown
+    _options?: unknown,
   ): void {}
   removeEventListener<K extends 'change'>(
     type: K,
     listener: (this: MediaQueryList, ev: MediaQueryListEventMap[K]) => any,
-    options?: boolean | EventListenerOptions | undefined
+    options?: boolean | EventListenerOptions | undefined,
   ): void;
   removeEventListener(
     type: string,
     listener: EventListenerOrEventListenerObject,
-    options?: boolean | EventListenerOptions | undefined
+    options?: boolean | EventListenerOptions | undefined,
   ): void;
   removeEventListener(
     _type: unknown,
     _listener: unknown,
-    _options?: unknown
+    _options?: unknown,
   ): void {}
   dispatchEvent(_event: Event): boolean {
     return true;
