@@ -2,6 +2,6 @@ export class MyTel {
   constructor(
     public area: string,
     public exchange: string,
-    public subscriber: string
+    public subscriber: string,
   ) {}
 }
