@@ -3,7 +3,7 @@ import { startWith } from 'rxjs/operators';
 
 export function conditionalStartWith<T>(
   condition: () => boolean,
-  cb: () => T
+  cb: () => T,
 ): OperatorFunction<unknown, T> {
   return function (source: Observable<any>) {
     return defer(() => {
